segmentation fault (enjoy that one)
unfortunately, this thing currently segfaults.almost certainly, this
is to do with the attempts to access the triangle ids in the half edge
constructor (we need them to be able to orientate the half edge wrt the
triangle). This was only a hacky temp solution so it needs some
thought (prehaps we instead pass in a position, 0,1, or 2 when a
halfedge is constructed? The index positions of the other vaules could
then be deduced). Still, the big problem with the original reason for
this branch is now fixed. Once we have tidied up a little and figured
out a method to verify if the mesh is contiguous, I can't think of any
other major problems with the implimentation, aside from triangles with
0 area/ inf cots (is this possible?)