huge (but partial) refactoring. C++ needs work
most of the C++ code is now neat. Still want to go through and
remove one or two functions entirely as it's now it's fairly clear that
vectorizing repetitive computations and passing them into the C++ code
in a consistent manner is bound to be less error prone and faster
(for one thing, the entire vec3 class which was always pretty shitty
can go for sure).

After this point it would make sense to improve all the code base,
start documentation, and figure out the cholesky Poisson problem.