PIL->Pillow, remove numpydoc, tidy mesh import
Remove the dependency on numpydoc as for now we will just be using
default docstring format. Changed Image import dep from PIL to Pillow -
the only place this is used is in the mesh importer, so the import has
been updated. Few other little fixes in this mesh file too (god bless
PyCharm)