moved align into a sensible struct, all broken
this is hopefully the last of the radical commits moving huge chunks of
pybug around. Now most things seem to be in a sensible place - anything
that is in major need of work (or maybe just doesn't need to be included
at all) is now in ./tosort. the align package is totally broken at this
point - goal this afternoon is to get it all sorted using the new
transform module.