The almightly Lucas-Kanade commit (sort of)
This is so that James can get on with AAMs. I haven't finished
the documentation yet, so will branch back off master now and
start making the changes. I'm also missing tests, as terrible as
that is. I'm working on it now. I've also committed this as one
massive commit to try and keep master clean.

Add a new alignment method (Lucas-Kanade) inside the align
package.