Flattenable->Vectorizable
Removes the 'flatten' terminology from the code, replacing it with
'vector' which is a little clearer to me. Vectorizable things now have
to methods:

- as_vector(self)
- from_vector(self, vector)

Where from_vector is now an instance method, and self is used to
populate missing state that is required (masks in images, trilists in
trimeshs etc).

Next we need to change transforms to use the same mechanism to be
consistant.