PARTIAL: Fix the LSResidual to work with images
Most things are broken - do not trust this commit. Committing
before I leave for the night. LSResidual tests are currently
passing but there is lots of dead code etc that I haven't pulled
out. It's worth noting that the order of the shearing had to be
swapped as well, which may be correct but I haven't fully
thought about it