NOTE: Tests changed? - Possible fix for Gradient Correlation
I've had to fix the x and y axes to match the cartesian image
coordinates, which might not be correct? The values output are
very close to the Matlab ones but not identical. These tests seem
particularly flaky due to the sign flipping around really small
values, so we are subject to floating point errors. But, the
overall errors are very close (10**-6) for the Yale Database Test,
which suggests this is OK. We may want to try come up with a
version that doesn't flip the axes, but the values were *slightly*
off when I did this, and so it made me nervous...