Remove matlab interpolator
The Matlab interpolator doesn't work with masked images, as
interpn expects gridded data. I have a sort of example that shows
using griddata but it doesn't yield the same results as interpn.
Therefore, I've taken out the concept of an interpolator and just
replaced it with a custom warp function. Therefore, if we wish to
add a matlab warping function we can do so and still easily swap
it in. Preferably we will use some c-code (possibly something that
Joan has) that can be used in both Python and Matlab identically.