Abstract out viewers for our shape classes
This means that their is a base pointcloud, trimesh and landmark
viewer that abstracts out dimensionality. These do not inherit
from Viewer because they are ONLY meant to abstract out the
dimensionality. The current chain of viewing may want to be
simplified -> as in you call PointCloudViewer and it returns you
the correct object from the constructor?

I also had to refactor pointcloud kwargs so that the correct
viewer gets passed down. Kwargs for viewers are a bit of a mess
at the moment