Fix warping for masked images
When warping gradients in to the frame of an intensity image
we have a problem that intensity images only accept one channel.

However, all we really care about is the mask. So, let's just use
that and build a blank image of the correct type and shape and
use update_from_vector