pseduoinverse_vector on Transforms fixing AAMs
This commit adds the ability to call a (potentially efficient)
pure-vector pseduoinverse on Transform objects. A naive version is
provided in AbstractTransform so immediately the function can be used.
An overrided version is provided for ModelDrivenTransform which is very
efficient. Currently just the ImageInverseCompositional LK object uses
this method. If proven to work it will be rolled out to all other LK
methods.