Rasterizer Makefile for OS X and Linux
This adds a Makefile that checks the OS and successfully compiles the
rasteriser for both OS X and Linux. IT IS NOT GOOD CODE (there is a hard
path stuck in there for god's sake) but the hope is anyone looking at it
will be able to figure out the differences between the two platforms in
a minimal manor.

Note that this assumes glfw3 has been brew installed (particularly
version 3.0.3) and is in the default Cellar location at /usr/local/