supress feature_type assets in tests
We currently have a small wrapper around every feature we build that
ensures that constrain_landmarks_to_bounds() is called after feature
computation in the fit framework. This breaks these tests. It may be
desirable in the future to revisit this with a better solution, but
for now this is holding up a huge improvement to Menpo - we can live
with the features being a little less discoverable in the short term.