Update some cython code to handle indices better
On Windows, the default Python 'integer' type is not a long,
but a long long. In general, this means we need to handle either
long or long long. For example, in the normals.pyx code, we use
fused_types to template the different integer types. We were
getting a failed test on windows due to not using fused_types
in the extract_patches method. However, fused_types don't play
well with the buffer interface.

Therefore, to make our code a bit more robust to very large
indices, I've changed from int/long to long long for indices.
This means taking more memory, but allows us to create really
massive images without error. If this is a real problem in terms
of memory then we can investigate proper templating with fused
types at a later date.

Note I have not update LBP and HOG with these changes as these
files need properly updating anyway.