Template the gradient method - fix Py_ssize_t
In Python, indices are typedef at compile time to use the
Py_ssize_t type. This was causing compilation issues on Windows.
More specifically, we should make these changes across the other
Cython modules. However, at the moment this commit just fixes
extract_patches (and updates normals.pyx).

I also templated the gradient method so that we can use double
or float transparently. Thus, I added tests and ensured that
the new gradient method is identical to the numpy method
for 2D images. I also fixed a bug on the branch in that method
and updated the numpy method to maintain the correct ordering.
This means I removed the fast2d gradient flag.