Large Change for Viewable objects
This lays a big foundation down for properly supporting viewing
objects across 2D and 3D given the new API change. Since 2D
viewing now has all of the parameters directly on the view method,
these parameters break 3D viewing. Therefore, we needed to either
remove the parameters, or subclass every object for different view
methods for 3D and 2D.

Instead, we created a pretty hair decorator that simulates
polymorphism, it will swap out the view method in a lazy manner,
so it should have no runtime overhead (after initial import
of the class). This decorator also cleverly allows autocomplete
to still work.

This commit also has a lot of other quite large changes to the
class hierarchy to make sure that this is all consistent.

Namely, PointGraph is now a PointCloud and a Graph and
the LandmarkViewable class was broken apart.

Also, LandmarkManager no longer has a view method.

I know this is a pretty ugly commit, but we were working on it
all day and since viewing isn't mission critical I think it is
OK to drop this all down at once.

TODO: Documentation of the 3D methods
      Update 3D methods to work with the new schema.