Fix to numpy>=1.9.1 and fix daisy
There was a regression between numpy 1.8.1, 1.9 and 1.9.1 whereby
1.9.0 added second order edges for gradients. 1.8.1 was first order
and then people complained, so in 1.9.1 they regressed back to
first order edges and added a kwarg. Unfortunately, we wrote
our daisy tests against the second order edges, so our tests
started failing when numpy moved to 1.9.1 as a dependancy.

Therefore, we rely on >=1.9.1 so that we can use the new kwarg
that fixes first order edges, and I changed the daisy tests
to reflect this.

Also, I made a mistake in the setup.py version ranges, the larger
ranges should not have had the = (equals to).