IGO and ES were wrong from gradient changes
Gradient is now ordered all the ys and then all the xs. Before
it was y and x interleaved. Therefore, we needed to update IGO
and ES to that they calculated correctly. They now return all
the ys and all the xs in order. For double IGO, it goes
all the ys, all the double ys, all the xs, all the double
xs.