Upgrade versioneer to PEP440 compat version strs
Although versioneer is 0.13+dev and not a released version,
because we ship with the versioning script inside menpo and it
isn't a dependancy, it doesn't really matter. This makes
our version strings PEP440 compatible which is a big win for
pre-release versions and shuts up pip when installing menpo
in development mode.