Allow warping of large images using batches
The TPS cdist function creates enormous symmetric matrices
for large images (since there are a lot of indices to sample
inside large images). Therefore, this commit allows an optional
kwarg batch_size for warping, whereby the points to warp get
passed in smaller batches. By default, this is not used.