Breaking change for mask warping
Given the new sampling behaviour, there is now a new definition
for the behaviour of warping masked images. It will **only**
warp if the target mask is True everywhere that the reference
mask wants to sample. Otherwise, it will throw an
OutOfMaskSampleError, which will fix problems for things like
building AAMs with masked Images.

If you want to sample the whole image, you as_unmasked.
If you want to warp the mask, use warp_to_shape on the mask, warp
the unmasked image, then create a masked image and set the
warped mask on the warped image.