Fix TriangleContainmentError
Missed the case whereby all the points are INSIDE the bounding
box, aka the landmarks ARE a box. This should set all points to
true.

Also, the bounds are inclusive on the top end, so we add +1 to
the max bounds to be correct. Added tests for these cases
and checked on master.