Add new point_in_pointcloud kwarg to constrain
This new kwarg determines how the point in pointcloud test is
performed. By default it is still PWA, but there is a new, fast
method provided by matplotlib that does a basic convex hull
check, where points on the boundary are counted as **outside**
the hull.

Also, the trilist argument is deprecated and raises a warning.

Finally, the point_in_pointcloud can actually take an
arbitrary callable for overriding the test.