New utility methods - scale/rotate about centre
This relies on #590.

Small amount of refactoring, but it mostly adds two new utility
methods, scale_about_centre and rotate_about_centre. These
perform the common operation of translating an object to the
origin, rotate/scale and then translate back. It also updates
a number of methods to use these functions since we actually
already performed this in 2/3 places such as Procrustes.