Refactor out bounding_box method
At the moment, in menpofit and menpodetect, we have need
of creating bounding boxes, and it is awkward right now
because there is now way to do it. You have to create a pointcloud
and then call bounding_box on it. This makes this a lot simpler
by exposing a bounding_box function globally from the shape
module