Fixes for new cyvlfeat
This should fix menpo for the newly updated cyvlfeat - which
should now be consistently Y, X instead of X, Y and also
returns the arrays (n_samples, n_features) which makes
them C-contiguous. The only thing is that it seems equivalent
to the transpose in Matlab, which may be a mistake or may
be by design... who knows.