Trying to fix the negative values and avoid segfault
The Matlab code crashes as well apparently - it is not good
at generating correct values. This maintains our previous
indxing logic, but changes the interpolation code to use
fractional distance from the center of the bins. This
stops negative values. However, we need to check the bins
are getting interpolated correctly due to the unusual way
that they were being interpolated previously.