Fix bug with warping boolean images
Seems that Cython does not natively support Numpy's bool
type and therefore I need to cast to uint8 in the Python
code. If someone works out how to get fused types to coerce
numpy bool properly, or Cython updates to support the bool
type, we can remove this cast. Note that this was happening
regardless as it was being cast inside the boolean image
constructor.