Add new _labeller decorator
This decorator handles the new workflow for labelling methods.

Labelling functions should return
a template pointcloud and a mapping dictionary (from labels to indices).
This decorator then takes that output and returns the correctly labelled
object depending on if a landmark group, pointcloud or numpy array is
passed.

If a LandmarkGroup was passed, a LandmarkGroup is
returned. This landmark group will contain specific labels and
these labels may refer to sub-pointclouds with specific connectivity
information.

If a PointCloud was passed, a PointCloud is returned. Only
the connectivity information is propagated to the pointcloud
(a subclass of PointCloud may be returned).