Update the ibug_face_68 landmarks to new scheme
Now uses the _labeller decorator. Also, renamed it to match
menpobench since I think @jabooth had decided this was the most
sensible naming scheme. Can be easily renamed! For example,
we may want to use the optional argument on _labeller to make
the group label key shorter!

Also, update the labeller method to grab the group label off of
the function itself. All the other methods are thus broken at
the moment, but this is a work in progress.