Refactoring labels (take parameter)
Change the _labeller decorator a bit, to take an extra kwarg
that can return the mapping dict, if a PointCloud was passed.
This is useful for reusing annotation schemes.

Also, make the wrapped methods take an argument to allow for
formats that reduce the number of annotations.

Also, simplify the output labels by just taking the back half
of the name scheme.

This was then applied to all of the label functions we have.