Add the GMRF model
This is a squashed commit that adds the GMRF model as @nontas
originally defined it. Here I am extracting out the parts
that @nontas and I had merged when we added my PCA changes.
This includes mostly GMRF stuff (pcacov, and graph methods)
but it also includes the gaussian ellipses changes.