Add LazyList - default importing is now Lazy
This adds the concept of an Immutable LazyList class. This
is essentially a very simple class whereby the elements are
callables that are assumed to do something *on index*. This is
very useful for importing because we just immediately return
a set of partial import methods that wrap the file paths that
need to be imported. The disadvantage is obviously that if the
file system changes after the list is constructed then the
import will fail. However, the low memory and instant access
(think random access) is a great benefit. To allow for backwards
compatibility there is a new as_generator argument that restores
the previous behaviour.

Finally, there is one major change - now, importers that return
lists of assets will not flatten the overall list - for example,
if you import a GIF, previously the frames would be yielded as
seperate images. Instead, the LazyList will not contain a list
that contains all the frames of the GIF. If this is a serious game
changer then we can add a helper generator that will flatten
the lists.