Add 2D landmark rasterize method to images
This method allows rasterizing of landmark groups onto an image.
Multiple landmark groups can be rasterized onto the same
image for efficiency. Currently two backends are provided, pillow
and matplotlib. The matplotlib backend is far more feature
complete, but slow! The pillow backend is much simpler but
an order of magnitude faster than the matplotlib backend.