WIP: BREAKING: Various release fixes (#687)
* Add test for set_boundary_pixels

* Fix some documentation in MaskedImage

* Fix set_boundary_pixels

Docs were incorrect, was not actually returning a copy and so now
it does. This is a breaking change!

* Refactor normalize method

Deprecate inplace methods and make normalize a feature. Add more
tests to ensure that everything works as expected.

* Gaussian/pyramids return copies

Add tests for pyramid shapes

* Add boolean image constrain tests

This is actually already covered by the constrain_mask tests
but at least this ensures that changing the return signature
will not throw an error.

* constrain_to_landmarks/pointcloud returns copy

BREAKING CHANGE: No longer in place - now returns a copy.

* Deprecate build_mask_around_landmarks

Replace with the more verbose but sensible name of
constrain_mask_to_patches_around_landmarks

* Deprecate set_masked_pixels in favour of from_vector

This is equivalent anyway so this method is just sugar that
is confusing.

* BREAKING: set_patches is no longer in-place

It now returns a copy of the image.

* Allow setting as_unmasked with iterable

Makes it easier to set colour images with colour backgrounds
by passing a list/tuple of floats for example

* Add rasterize_landmarks

Exposes the rasterization method

* Add bounds() method to image

The bounds are defined by ((0, 0), (height - 1), (width - 1)).
This is so that constrain_landmarks_to_bounds can use
these bounds. This makes sense when you consider the bounds as
being inclusive of the set of integer indices it is possible to
index into the image with.

* BREAKING: Add the constrain_to_bounds method to pcloud

Deprecate the constrain_landmarks_to_bounds method.
Change has_landmarks_outside_bounds to a method

* Fix typos pointed out by @nontas

* Add init_from_pointcloud

Needed overloads for each of the image types - was easier
just to duplicate the very small amount of code than try and
make some hierarchy of functions for the sake of 3 lines.

* Deprecate normalize methods on image

Add other normalize methods to features including norm, std and
var.

* Standardize on normalize vs normalise internally

Prefer american spelling

* Add raises info to normalize feature

* return_transform to init_from_pointcloud

* Documentation fixes

* Deprecate MaskedImage normalize

* More docs fixes

* Make normalize an image feature

now when masked image normalize is deprecated we can actually rely
on the feature