Allow passing exporter kwargs - allow path only exporters
This makes two changes - one is that we now support path only
exporters (via a private method). This has been dogfooded inside
and was really intended to be used with Menpo3D since the VTK
exporters can't write directly buffers. Actually, we already
had this paradigm with videos so this just formalises that.

It also allows passing extra kwargs to exporters in the same
manner that videos already can.