Fix pip install by properly including source files
I misunderstood how the packaging worked and naively used
the package-data key to add the header and cpp files to the menpo
install directories. However, this only works for cpp and header
files are that inside **packages**. So it doesn't include, for
example, headers in places such as the cpp/ folders we use.
Thus, we revert back to using the MANIFEST.in and remove
the package-data as we never need to rebuild menpo when
installed in site-packages directly.