Update landmark resolver logic to support dicts
Now the resolver/landmark import logic returns dictionaries to
allow for importing LJSON files that map much more closely
to our landmark managers. This means a breaking change but it
is the only way to allow for supporting this. To try and make
the change less painful, the import_landmark_file gains a new
kwarg that allows to choose a single group to return (to more
closely match the old API)