Fall-back to querying numpy for inc dir
We changed to check site_packages for numpy's inc dir to avoid having to
import numpy at setup.py invocation. That's great, but in some
environments that means we will not find numpy correctly.

This doesn't change initial behavior, but at the point where we used to
give up and fail to install it will now attempt to import numpy to find
the inc path. If numpy fails to import then we back out as before.