Fix rasterization test for Matplotlib 2.0 and 1.5
Looks like the style changes in Matplotlib 2.0 cause weird
issues so I just use the context manager to ensure
that the 1.5.x classic style is used only for the rasterization.
In Matplotlib 2.x it's really weird that just trying to rasterize
a line always returns a line that is too short? Might
be worth raising on the mailing list or something as I'm at a loss