Fix boundary_tri_index
As far as I can tell this was actually broken because it
didn't properly account for repeated but "swapped" edges meaning
that all edges were counted as unique. From manual testing it
appears to be that now this is actually correct. It also meant
changing the ordering of the returned edges to make iterating
easier.