Fix boundary_tri_index (#832)
* Fix boundary_tri_index

As far as I can tell this was actually broken because it
didn't properly account for repeated but "swapped" edges meaning
that all edges were counted as unique. From manual testing it
appears to be that now this is actually correct. It also meant
changing the ordering of the returned edges to make iterating
easier.

* Fix trimesh unit tests due to edge reordering

The ground truth values are permuted based on the change in
ordering. Also use np.testing.assert_allclose

* Add test for boundary_tri_index()

Shows that this method works as expected