Change to pure Python patch extraction
This is slower than the Cython implementation but reduces
the maintenance of the package as a whole by removing the need
for C-extension compilation. It also adds the ability to
interpolate patches with strategies other than nearest neighbour
aka slicing. For example, bilinear interpolation can now be
used. Note this also changes how boundaries are handled as
patches are now always returned and filled with a constant
value rather than being truncated.

For reference, on the masked Takeo image - extracting patches
around the landmarks previously took ~300us on my 2013 Macbook Pro
and now takes ~550us - so close to twice as slow.