Make verifying the mask optional
We perform a helpful operation when we sample a masked image and
check that the sampling only occurs in valid pixel regions. This
is a nice check but it makes things much slower for simpler
operations like resizing. Therefore, add a flag to enable this
behaviour and disable it by default.