Simplify the license to an unmodified BSD 3-Clause
The original license was in the style of the BSD 3-Clause
license used by the Scipy project. This has caused some
confusion and thus the license is being simplified back to
a totally unmodified BSD 3-Clause license. Note that this
is not intended as a change of license and agreement from the
majority of the original contributors has been acquired
at https://github.com/menpo/menpo/issues/840.

Nevertheless - this is done for transparency - no license
change is made here - simply the license file is modified.